python310Packages.cython: Force regeneration of generated code in sdists
Lots of python packages seem to ship their source distributions with
generated code. In recent cython versions there is a way to force a
regeneration of these files and using it sounds generally sounds like a
good thing.

If it turns out, that this causes problems on individual packages,
setting `dontForceRegenCython` will prevent the regeneration.

Added in 3.0.0a11 and backported to 2.9.34, so both cython versions we
ship right now support this flag.