flake.nix: provide nixpkgs used to build nixos to modules
This allows NixOS modules to take `inputs` as an extra argument. This
may be useful for feeding version information into other config.

For example, I use this to inform a module which exposes a prometheus
metric on the date of the nixpkgs version used to build the
system (slightly abridged):

```
{ pkgs, config, inputs, ... }: {
  services.nginx.enable = true;
  services.nginx.virtualHosts."prometheus".root = pkgs.writeTextDir "/metrics" ''
    # HELP nixpkgs_date The timestamp of the last modification on the nixpkgs flake input used to build the system.
    # TYPE nixpkgs_date gauge
    nixpkgs_date ${toString inputs.nixpkgs.sourceInfo.lastModified}
  '';
}
```

This is mostly a convenience feature. Downstream flakes were already
able to set specialArgs.inputs themselves, but this makes
obtaining correct nixpkgs version information even in a flake which
uses multiple different nixpkgs versions a little more elegant.