pihole: init at 6.0.6
Which is the core admin script for pihole, which modifies the pihole-FTL and
gravity (DNS settings) database.

It requires quite a lot of patching to remove most of the hard-coded absolute
paths, and add the many dependencies the script has.

Patches also remove numerous unnecessary/unsupported operations from the script,
including reinstalling, uninstalling, and updating pihole, pihole-FTL, and
pihole-web.

Some scripts also include the installation scripts for some utilities, so these
utils need to be copied over to the files which are included in the Nix package.