nixos/pihole-ftl: init
Add a module for pihole-ftl, which allows declaratively defining the
setupVars.conf and pihole-FTL.conf configuration files.

Also provide options for adlists to use, which can be added through the pihole
script (packaged as "pihole"). Other state such as clients and groups require
complex database operations, which is normally performed by the pihole admin
webapp (packaged as "pihole-admin").

Extend the dnsmasq module to avoid duplication, since pihole-ftl is a soft-fork
of dnsmasq which maintains compatibility.

Provide the pihole script in `environment.systemPackages` so pihole-ftl can be
easily administrated.