dante: disable getaddrinfo() checks if cross
Prior to this commit, dante's configure-time getaddrinfo() checks
were disabled only if stdenv.hostPlatform.isMips64.

These checks must also be disabled if the buildPlatform cannot
execute hostPlatform binaries.  This commit factors out the control
of this disablement as a flag and adds an additional situation in
which that flag is enabled.