etlegacy: init at 2.81.1
See packaging request: etlegacy - fixes NixOS#101187.

This change has been updated in light of feedback from @teto
to use a more conventional structure consisting of a single
stdenv.mkDerivation, with the wrapper moved into postInstall.

fetchGit was substituted for fetchFromGitHub.

The 'fakeGit' script was greatly simplified to simply echo
${version}.

meta.mainProgram was also added, as the filename of the main
binary varies according to the architecture, and is referenced
in the XDG .desktop files. This also avoids collisons as it's
not overly uncommon for people to have both 32- and 64-bit
versions of this game installed at the same time.

meta.license was updated to include the cc-by-nc-sa-30 license,
which some resources use.

Finally the configuration was passed through the alejandra
nix formatter.