mongosh: fix hash now that all git deps are fetched
This FOD hash changed becuase it the fetcher was not downloading dev
dependencies from https://github.com/mongodb-js/saslprep/tree/v1.0.4 (a
transitive dependency) and so since including git dep lockfiles is the
purpose of these changes this change is intended and the hash should be
updated.

Generally a package was unable to build without these so I did not
expect this to be a breaking change (since there would be no existing
FODs), but mongosh seems to have been an exception. I think just
changing the hash here is more appropriate for this exception rather
than working in machinery for supporting old and new FODs (since it's
rather unlikely these prefetch-npm-deps changes will break downstream
code as it should not have worked in the first place if it needed this).