nixosTests.cage: fix OCR properly
The underlying problem with OCR in this test has been that the only
font installed was DejaVu Sans, a proportional font, which xterm would
try to render as monospace.  This produced very broken looking text,
which the OCR understandably had trouble with.  With an actual
monospace font installed, there are no more problems and we don't need
the hacks.