noice-nvim: Remove nvim-notify from noice-nvim dependencies
*!This will change behavior for the end user of noice as they will have to add notify plugin manually!*

Notify plugin is no longer a hard dependency for noice plugin and there is no need to ship noice with it. 

In addition to that some plugins will perform auto discovery of plugins that are on the runtime path and in case they find notify they might want to use it.
There is no easy way to remove nvim-notify from the runtime path without doing an override. 

https://github.com/folke/noice.nvim#%EF%B8%8F-requirements