apple-framework-System: fix System.tbd dangling symbolic link
Contrarily to the other frameworks, System framework's TBD file
is a symlink pointing to `${MacOSX-SDK}/usr/lib/libSystem.B.tbd`.

This produces an error when using the framework, as:

  1. The original file is not copied into the output directory
  2. Even if it was copied, the relative path wouldn't match

Resulting in the symlink being broken and the linker failing when
trying to link `-framework System`.

The fix applied consists in replacing the symbolic link with the
actual file, as this is easier than fixing the link and doesn't
seem to produce any side effects.