fcft: add zlib license to list
fcft bundles nanosvg in-tree and by default builds it into the library.
nanosvg is licensed under the zlib license.

Following the discussion on fnott [0] (which does the exact same), it
was decided that the license meta-attr should reflect this situation by
including the zlib license.

There are other precedents as well, like the NixOS [1] and Debian [2]
package for fuzzel.

Thus, the same should be applied to fcft as well.

[0] https://github.com/NixOS/nixpkgs/pull/225224
[1] https://github.com/NixOS/nixpkgs/blob/master/pkgs/applications/misc/fuzzel/default.nix
[2] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=982145#74

Signed-off-by: Christoph Heiss <christoph@c8h4.io>