qt6: drop cmake patch used for fixing cmake file generation
The "multiple-outputs.sh" setup hook moves "include", "lib/cmake" and
other folders into the "dev" output if it exists, thus breaking the
invariants expected by the qt build system and we used to patch cmake to
fixup the generated cmake files. In a series of changes to rework qt
packaging, we are now setting "moveToDev" to false to suppress that
behavior, and the cmake patch is no longer required.