openrct2: Refactor to allow version override
Since OpenRCT2 requires server and client versions to be equal, it is
desirable to be able to override the openrct2 pkg version number. By
moving the version numbers and checksums to the function parameters we
make them easier to override by consumers.

The logic that checks if the correct dependency versions are supplied
now supports checking dependencies that are only required by some
openrct2 versions. Additionally it includes the version numbers in the
error messages making it more convenient to iteratively specify your
desired package.

Tested by building (and verifying) openrct2 versions 0.4.1, 0.4.2, 0.4.3
(I think) and 0.4.4.