submitting-changes.chapter.md: explain *why* we have three branches
The manual does an okay job of explaining the rules for each of the
three development branches, but really doesn't give any intuition as
to why there are three (why not four? or two?) or how we got where
we are today.

This commit attempts to fix that, by explaining that there is one
branch that allows mass-rebuild commits, and it has a fast-building
branch both upstream and downstream of it (from the perspective of
automated merges).

I have also removed the term "stabilization" from the arc labels.
This vague term is not defined anywhere, and does communicate any
useful information without a longer explanation.  Therefore it is
not appropriate for use in a diagram.

Co-authored-by: Valentin Gagarin <valentin.gagarin@tweag.io>