nim: propagate Security framework to all nim packages on darwin
Many nim packages regressed on darwin after 31254120db53fe07d9f1f05fa8ff41af8feaac03, which upgrade nim to 1.6.12. The cause of the breakage is that the nim standard library was changed to link against the Security framework when using the system rng on darwin. Notably not *all* nim packages were broken by this, since packages that don't use this part of the standard library don't need the framework to build.

This commit adds the Security framework to the depsTargetTargetPropagated for the compiler wrapper, since the standard library isn't fully functional without it.