octavePackages.control: update 3.5.1 -> 3.5.2
Also update to use the GitHub version of control, as development has
moved there from SourceForge and 3.5.1 is no longer obtainable from
SourceForge.

Add autoreconfHook so that config.h header is generated for control by
control.
This seems to be a regression, as 3.5.1 and earlier should have been
building fine without it?