doc: pass nixpkgs-revision instead of nixpkgs
The `doc` expressions take both a `pkgs` packageset as well as a
`nixpkgs` argument whose purpose is unclear.

I traced all the usage of `nixpkgs` and it appears to be simply "any
attrset with a `revision` attribute".  Let's replace this argument
with `nixpkgs-revision` so people know what its purpose is.

I did not change the arguments to `doc/default.nix` because it might
be considered an exposed public interface (does Hydra
`nix-instantiate` it directly?).  We should give it an entry in
all-packages.nix and access it that way.