yosys-symbiflow: 2023.02.08 -> 1.20230425
The plugin now has a tagged version that is based on the date.
Previously, we had to use a random commit and manually choose the
relevant date. Now, we use the official tag as a version here
and to fetch from git.

While at it: fix the tests - somewhere over the course of the
lasts year, the tests stopped being run. Fixed now.