clipmon: init at unstable-2022-08-28
Under Wayland, an application itself is responsible for "serving" the
current clipboard contents. As soon as that happens, `clipmon` claims
the clipboard and copies the contents from the application[1].

As a result, clipboard contents are still available after closing the
application where something was copied from.

Additionally, I implemented a small VM test to check the behavior when
using Sway like this:

* Open a terminal and write something to the clipboard using `wl-copy -f`.
  That way the clipboard is held by a sub-process of the terminal (which
  simulates a Ctrl-Shift-C in `foot`, not using `-f` would cause
  `wl-copy` to fork itself to keep the contents after closing the
  terminal which wouldn't happen when copying by e.g. mouse).
* Ensure in a second terminal that the clipboard contents can be
  accessed.
* Close both terminals. Without `clipmon`, the selection would be lost
  now.
* Open another terminal and run `wl-paste` to ensure that the selection
  still exists (because of `clipmon`).

[1] https://git.sr.ht/~whynothugo/clipmon#design