honggfuzz: install libraries, too
honggfuzz embeds its .a libraries *inside* of the hfuzz-cc executable,
extracts them when you invoke it, and links the .o file against them.
But that doesn't help if you wanted to bind it to other languages such
as Rust, for example. If you're using honggfuzz-rs, it embeds this
automatically too, but if you're doing it yourself, you need these
libraries to get anywhere.

Signed-off-by: Austin Seipp <aseipp@pobox.com>