systemd: disable libbpf if compiler-rt unsupported
If we can't build compiler-rt, we can't have a clang for that
platform.  Example affected architecture: s390, which is useful for
testing assumptions in Nixpkgs because it's a Linux architecture that
we can't emulate.

I've written buildPackages.targetPackages.llvmPackages even though
it's the same thing as llvmPackages because of the comment in this
file that warns people against relying on splicing for llvmPackages.
Taking llvmPackages as an input directly would make it easier for
people to make that mistake without seeing the comment.