vim: Harden build against ex and other tools being missing
Those tools are often used as build-time dependency. Some deficiencies
in the upstream makefiles makes it possible that they will not be
installed, even though the `make install` call was successful.

This *could* make the `vim` build fail where it wouldn't have failed
previously, but it will now fail in a correct manner instead of being
incorrectly successful.

The previous patch (actually using -j1 in install phase) should make
this change redundant, but let's not pretend new bugs are never going to
happen!