chromium: (cross): use cc-wrapper-clang instead of bare clang
Prior to this commit chromium was writing the bare, unwrapped
`clang` path into `gnFlags`.  When called without the wrapper, clang
doesn't know that the linker for cross compilation isn't called
`lld` anymore -- it is called `${targetPrefix}-lld`.

This commit fixes that problem by referencing the cc-wrapper-ified
clang which is part of `llvmPackages.stdenv`.