chromium: (cross) strip aarch64-linux-gnu- toolprefix only for native builds
Our existing chromium expression strips the `aarch64-linux-gnu-`
prefix off of toolchain names (like `aarch64-linux-gnu-gcc`) for all
`Aarch64` builds.  But when we are doing a cross compile *to* an
`Aarch64` hostPlatform we need to leave those prefixes in place.

This commit makes the prefix-stripping conditional on the build
being a native build.