wrapGAppsHook: don't add superfluous inputs
As noted in the comments, there's no reason for wrapGAppsHook to
propagate gtk3 at all, or librsvg as a native build input.  This also
means we can get rid of wrapGAppsHook4, as it's now identical to
wrapGAppsHook.

It's finally time to do this, because I noticed the extra gtk build
input was breaking cross compilation of gcr_4.