pyqt5: add enableVerbose?true
The build logs from PyQT5 are quite verbose; it's nice to trim them
down so all you see is the errors.  This commit adds an option which
does that.  It also adds "-w" to `NIX_CFLAGS_COMPILE` in non-verbose
builds, because newer compilers keep piling on the warnings and Qt5
isn't getting updates.