linux/kernel/generic.nix: assert if constraints would be ignored
There appears to be a real mess with the naming of the attribute
which contains structured extra configuration:

- Nixpkgs' linux kernel expression uses structuredExtraConfig.
- NixOS uses extraStructuredConfig.
- Nixpkgs' kernelPatches use extraStructuredConfig.

At the moment, if the user uses the wrong convention the
configuration constraints are silently dropped.  Let's `assert`
instead so they at least know that their request isn't being
respected.  We can try to harmonize the naming later.