system.boot.kernel: warn about requiredKernelConfig limitations
I struggled a lot before realizing that
system.boot.kernel.requiredKernelConfig was not only being ignored
(due to kernel?features) but wasn't even really checking my
requirements against the true kernel configuration.

This commit updates the option documentation to warn others of the
limitations I discovered.

Long-term we should probably either remove this option or else have
it check the requirements against the true, final `.config` that was
used to build the kernel.