mosquitto: use multiple outputs
Most importantly, separate the library from the executables. When the
library is used as a standalone MQTT client library, then we don't
want to pull in, e.g., the broker daemon.

Conveniently, the library output does not have a runtime dependency on
systemd so it becomes unnecessary to use `withSystemd = false` to keep
the closure size down when only wanting to use the library.