virt: Don't insist on kvm support
We want to *build* VM images without considering the build-host's
capabilities.

KVM is never a requirement; VMs runs fine without KVM - albeit slower.
This puts the onus on making KVM an opt-in capability, rather than a
necessary requirement.

Besides, QEMU is perfectly able to fallback from KVM to TCG JIT but that
can never happen if nix-daemon prevents it from ever trying.

In the current situation:
 - one *has to* modify nix.conf on Non-NixOS
 - AND have a CPU with virtualization extensions
 - AND have the {amd, intel}-kvm kernel module loaded
 - AND have correct permissions to /dev/kvm
 - and that is too much to insist on - certainly for *building* images.