pythonPackages.hasPythonModule: Deprecate calling with non-derivation
This adds a warning when hasPythonModule is called with a value that is not a
derivation.

This happens if the `buildEnv.override { extraLibs = [ ... ]; }` list includes
non-derivations, such as nested lists (and so also affects any callers of
this, like `python.withPackages`).

This provides a warning about an easy mistake to make, where a list of Python
module derivations is nested inside the top-level list.  This case currently
passes silently and just doesn't add any of the nested derivations to the
environment.