nixos/netboot: use initialRamdisk and prepend the nix store squashfs
Previously, we had `netbootRamdisk` pointing to the "final ramdisk", but
that's somewhat wrong, we have no way to say "this is the good ramdisk
to use for boot" purpose across nixpkgs.

Moreover, we use bootStage2 rather than the whole top level for the
store contents, that's because we have to break the infrec loop.

We do not need stage 1 stuff at that moment, breaking the loop
correctly.

Therefore, we just deprecate it, prepend to the existing netboot, making
us more compatible with systemd stage 1 for example.