clipman: 1.6.1 -> 1.6.2, re-add myself as maintainer, install man-pages
ChangeLog: https://github.com/yory8/clipman/releases/tag/v1.6.2

While it's not the ideal tool, it's the closest one to what I want to
have (clipboard contents being retained after closing a terminal on sway),
so I guess I'll have to stick to it for now. And since I already
packaged it, I decided to just become a maintainer again.

There are no functional changes in this release, however the man-page
got improved. Adapted the package to also install it.