kicad.base: handle bundle fixups manually on macOS
KiCad ~7 added these changes to support Apple Silicon. As detailed
within, `fixup_bundle` fails for us because it runs `otool` on paths
that do not exist and `IGNORE_ITEMS` does not seem to let us opt out of
this.

This is suboptimal (we'll need to keep our fixups in sync with those in
the KiCad build and missing fixups don't always manifest as loud startup
errors –– some only trigger when loading a plugin, etc.) but it _does_
work.