gnuplot_qt: fix build failure
Recently, wrapQtAppsHook was switched from shell to binary wrapper.
The binary wrapper does not support --run, which is used by gnuplot.
When gnuplot is built with Qt terminal, it "inherits" the wrong
wrapping function from Qt. In this commit we replace the wrapping
function back to the one generating a shell wrapper.

Fixes #229425.